refactor(site): move AgentChatPageView to correct directory
