+++ site/src/pages/AgentsPage/AgentChatPageView.stories.tsx
@@ -10,14 +10,14 @@ import {
 	withAuthProvider,
 	withDashboardProvider,
 } from "#/testHelpers/storybook";
-import type { ChatDetailError } from "../utils/usageLimitMessage";
+import type { ChatDetailError } from "./utils/usageLimitMessage";
 import {
 	AgentChatPageLoadingView,
 	AgentChatPageNotFoundView,
 	AgentChatPageView,
 } from "./AgentChatPageView";
-import { createChatStore } from "./ChatConversation/chatStore";
-import type { ModelSelectorOption } from "./ChatElements";
+import { createChatStore } from "./components/ChatConversation/chatStore";
+import type { ModelSelectorOption } from "./components/ChatElements";
 
 // ---------------------------------------------------------------------------
 // Shared constants & helpers
